Fix a bug with empty actions column when browsing data
When no action is allowed (eg. when GuiControl removes them) on the
browsing page (display.php), the "action" header was still showed, but
not its data column, leading to a difference between the number of header
columns and data columns.

This bug was raised by Julien Rouhaud (rjuju). This patch is based on
his work (some cleanup).