Instanciate $plugin_manager from login.php script if needed
This answers the question in commit message:
7d89d7cca40b5a8c3ac1874436630993f518410c. When the script require a
database connexion and the user must login, or in some cases of some
other related errors, we ar eincluding the login.php script and exit
immediatly. The login.php script requires the $plugin_manager, but it
is instanciated AFTER the connection codeblock (see commit 7d89d7cc),
leading to a FATAL error.