At last, the long awaited patch for php7 support.
This commit is focused strictly on the class/constructor changes. This has
primarily been tested on php 7.1 and postgres 11, with all tests passing though
there are some spurious warnings; I have fixes for them but will add them in
seperate commits.

Note that the code here is my own, but I did look at patches from
@gabrielhomsi and @w1ldzer0 for some sanity checking. Any bugs or missing
items are on me.