Fix for SF BUG #470 Can't list/see/edit sequences owned by group
While I have implemented this fix against Postgres 12, as far as I can tell,
this bug existed since 7.x, so I'm adding it to all versions back to 7.4
(when we did the class re-org). This could probably use more testing on older
versions though.

Thanks to Carlos Martinez for the bug report and reproducible test case!