Overhaul sequence properties page to better handle different permissions.
Whilst fixing the sequence group viewing bug, noticed that sequence properties
in general would throw an error if you did not have appropriate permissions.
While this is similar to permission denied errors when looking directly at
a sequence, there is other information to be exposed which is accesible
through pg_sequence and pg_sequences. The change implements a method for
determining if a user can access the sequence directly (ie. select * from seq)
and if so, include that information in the output. If not, we null those
columns.

This also fixes cases where last_value information is not visible due to a
sequence not haveing been "read" yet. ie. after sequence creation, or anytime
a sequence is altered, and nextval/setval hasn't been called.

This problem seems to go back to at least postgres 9, so I've mainlined the fix.
It might break on older unsupported versions of postgres; if someone wants to
fix that well... patches welcome.