Support truncation of mulitbyte srings
This fixes https://sourceforge.net/p/phppgadmin/bugs/422/ :
substr truncates on a byte-level, sometimes within a multi-byte character. This resulted in the whole string sometime not being displayed. See the original bug report for a way to reproduce.

Please note that this requires php-mbstring to be installed. This is usually the case, but the dependency should be described in the INSTALL file or something.