Fixes removal of oids in Postgres 12.
The primary goals here are to make things work for Postgres 12, but we also
want to keep support for oids in older versions just in case. There is also a
side wrinkle that we used to leverage oids for editing rows in tables without
primary / unique keys; I've left that plumbing in place for now, but the for
now we've lost that functionality.

Fixes sparked by PR #70, though this is not based on that PR.